Resume_Application_Tracking_System(ATS)_Using_Gemini_ProVision/requirements.txt
streamlit
langchain-community
google-generativeai
python-dotenv
pdf2image
ibm_watson_machine_learning